+++ packages/rx_bloc_cli/lib/src/models/configurations/feature_configuration.dart
@@ -4,6 +4,7 @@ class FeatureConfiguration {
   FeatureConfiguration({
     required this.changeLanguageEnabled,
     required this.counterEnabled,
+    required this.qrScannerEnabled,
     required this.widgetToolkitEnabled,
     required this.analyticsEnabled,
     required this.pushNotificationsEnabled,
@@ -34,6 +35,9 @@ class FeatureConfiguration {
 
   /// Counter
   final bool counterEnabled;
+
+  /// QR scanner
+  final bool qrScannerEnabled;
 
   /// Widget toolkit
   final bool widgetToolkitEnabled;