+++ packages/rx_bloc_cli/mason_templates/bricks/feature_profile/__brick__/lib/l10n/network_lookup/profile_lookup.dart
@@ -21,8 +21,6 @@ class AppI18nProfileLookup extends I18nFeatureProfileLookup {
           bundledI18n = I18nFeatureProfile(I18nFeatureProfileLookup_bg());
           break;
         case 'en':
-          bundledI18n = I18nFeatureProfile(I18nFeatureProfileLookup_en());
-          break;
         default:
           bundledI18n = I18nFeatureProfile(I18nFeatureProfileLookup_en());
           break;