+++ packages/rx_bloc_cli/mason_templates/bricks/feature_qr_scanner/__brick__/lib/l10n/network_lookup/qr_code_lookup.dart
@@ -1,28 +1,28 @@
-{{> licence.dart}}
+{{> licence.dart }}
 
 import '../../assets.dart';
 import '../{{project_name}}_app_i18n.dart';
 import 'util.dart';
 
-class AppI18nLibDevMenuLookup extends I18nLibDevMenuLookup {
+class AppI18nQrLookup extends I18nFeatureQrLookup {
   @override
   String getString(String key, [Map<String, String>? placeholders]) {
-    return getFromNetwork('lib_devmenu__', key, placeholders) ??
+    return getFromNetwork('qrCode__', key, placeholders) ??
         getFromBundle(AppI18n.locale?.languageCode)
             .getString(key, placeholders)!;
   }
 
-  I18nLibDevMenu? bundledI18n;
+  I18nFeatureQr? bundledI18n;
 
-  I18nLibDevMenu getFromBundle(String? languageCode) {
+  I18nFeatureQr getFromBundle(String? languageCode) {
     if (bundledI18n == null) {
       switch (I18n.currentLocale?.languageCode) {
         case 'bg':
-          bundledI18n = I18nLibDevMenu(I18nLibDevMenuLookup_bg());
+          bundledI18n = I18nFeatureQr(I18nFeatureQrLookup_bg());
           break;
         case 'en':
         default:
-          bundledI18n = I18nLibDevMenu(I18nLibDevMenuLookup_en());
+          bundledI18n = I18nFeatureQr(I18nFeatureQrLookup_en());
           break;
       }
     }