+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/l10n/network_lookup/field_lookup.dart
@@ -21,8 +21,6 @@ class AppI18nFieldLookup extends I18nFieldLookup {
           bundledI18n = I18nField(I18nFieldLookup_bg());
           break;
         case 'en':
-          bundledI18n = I18nField(I18nFieldLookup_en());
-          break;
         default:
           bundledI18n = I18nField(I18nFieldLookup_en());
           break;