+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/l10n/network_lookup/notifications_lookup.dart
@@ -22,9 +22,6 @@ class AppI18nNotificationsLookup extends I18nFeatureNotificationsLookup {
               I18nFeatureNotifications(I18nFeatureNotificationsLookup_bg());
           break;
         case 'en':
-          bundledI18n =
-              I18nFeatureNotifications(I18nFeatureNotificationsLookup_en());
-          break;
         default:
           bundledI18n =
               I18nFeatureNotifications(I18nFeatureNotificationsLookup_en());