+++ packages/rx_bloc_cli/test/models/configurations/feature_configuration_test.dart
@@ -21,6 +21,7 @@ void main() {
         cicdEnabled: false,
         cicdGithubEnabled: false,
         cicdCodemagicEnabled: false,
+        qrScannerEnabled: false,
         profileEnabled: false,
       );
       expect(sut.usesFirebase, isTrue);
@@ -38,6 +39,7 @@ void main() {
         cicdEnabled: false,
         cicdGithubEnabled: false,
         cicdCodemagicEnabled: false,
+        qrScannerEnabled: false,
         profileEnabled: false,
       );
       expect(sut.usesFirebase, isTrue);
@@ -55,6 +57,7 @@ void main() {
         cicdEnabled: false,
         cicdGithubEnabled: false,
         cicdCodemagicEnabled: false,
+        qrScannerEnabled: false,
         profileEnabled: false,
       );
       expect(sut.usesFirebase, isTrue);
@@ -72,6 +75,7 @@ void main() {
         cicdEnabled: false,
         cicdGithubEnabled: false,
         cicdCodemagicEnabled: false,
+        qrScannerEnabled: false,
         profileEnabled: false,
       );
       expect(sut.usesFirebase, isFalse);