+++ packages/rx_bloc_cli/test/models/generator_arguments_test.dart
@@ -41,6 +41,7 @@ void main() {
       cicdEnabled: true,
       cicdGithubEnabled: true,
       cicdCodemagicEnabled: true,
+      qrScannerEnabled: true,
       profileEnabled: true,
     );
     sut = GeneratorArguments(
@@ -91,6 +92,8 @@ void main() {
           equals(featureConfiguration.cicdGithubEnabled));
       expect(sut.cicdCodemagicEnabled,
           equals(featureConfiguration.cicdCodemagicEnabled));
+      expect(
+          sut.qrScannerEnabled, equals(featureConfiguration.qrScannerEnabled));
       expect(sut.profileEnabled, equals(featureConfiguration.profileEnabled));
     });
   });